backend/requirements.txt
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
sqlalchemy>=2.0.25,<3.0
pydantic-settings>=2.1.0,<3.0
python-jose[cryptography]>=3.3.0,<4.0
bcrypt>=4.0.1,<5.0.0
python-multipart>=0.0.6,<1.0
slowapi>=0.1.9,<1.0
